Fix travis ci build
