+++ test/com/facebook/buck/graph/transformation/BUCK
@@ -1,11 +1,36 @@
 load("//tools/build_rules:java_rules.bzl", "standard_java_test")
 
-standard_java_test(
-    name = "transformation",
-    deps = [
+java_library(
+    name = "childrenadder",
+    srcs = glob(
+        ["*.java"],
+        exclude = ["*Test.java"],
+    ),
+    exported_deps = [
         "//src/com/facebook/buck/graph/transformation:transformation",
         "//src/com/facebook/buck/util/types:types",
         "//third-party/java/guava:guava",
+    ],
+)
+
+java_test(
+    name = "transformation_unit",
+    srcs = glob(
+        ["*Test.java"],
+        exclude = ["DefaultAsyncTransformationEngineStackUseTest.java"],
+    ),
+    deps = [
+        ":childrenadder",
+        "//third-party/java/junit:junit",
+    ],
+)
+
+java_test(
+    name = "transformation_stackuse",
+    srcs = ["DefaultAsyncTransformationEngineStackUseTest.java"],
+    vm_args = ["-Xss200k"],  #explicitly set a small stack for testing stack use
+    deps = [
+        ":childrenadder",
         "//third-party/java/junit:junit",
     ],
 )
